Separate logging into separate files.
